修改建表语句

+++ src/main/java/com/hubu/aspirin/model/entity/Course.java
@@ -45,7 +45,7 @@ public class Course implements Serializable {
     private Integer period;
 
     @ApiModelProperty(value = "学分")
-    private BigDecimal credit;
+    private Float credit;
 
 
 }
